chore: split MessageReferenceList and MessageDocument into its own modules
